[REFACTOR] unifying the TableInfo class
Unifying the 2 identical TableInfo class in parsing_csv.py and parsing_xlsx.py
