+++ rag/parsing/processor.py
@@ -1,4 +1,9 @@
 from abc import ABC, abstractmethod
+from llama_index.core.bridge.pydantic import BaseModel, Field
+
+class TableInfo(BaseModel):
+    table_name: str = Field(..., description="Unique table name (must use underscores and NO spaces)")
+    table_summary: str = Field(..., description="Short, concise summary/caption of the table")
 
 # Abstract base class
 class FileProcessor(ABC):
